Commit Paolo: Moved Optimizers.py and changed readme.md
