+++ README.md
@@ -1,1 +1,48 @@
-Machine_Learning_Project
+# Machine Learning Project Documentation
+<p>
+The following file contains the documentation regarding the classes created for the ML Project 2021/2022.
+</p>
+
+<p>
+<h2>
+Classes documentation list:
+</h2>
+<h3>
+
+1. [Main.py](.\docs\mainDoc.md)
+2. [Model.py](.\docs\ModelDoc.md)
+3. [Model_selection.py](.\model_selectionDoc.md)
+4. [Layer.py](.\docs\layerDoc.md)
+5. [Optimizer.py](.\docs\OptimizersDoc.md)
+6. [Metrics.py](.\docs\metricsDoc.md)
+7. [LoadCVSData.py](.\docs\loadCSVDataDoc.md)
+
+</h3>
+
+</p>
+
+<p>
+<h2>
+ Quick start:
+</h2>
+
+The following section of the document is a guide to install and clone this project.
+
+
+</p>
+
+
+<p>
+<h2>
+ Software used
+</h2>
+
+To organize this project development and to make a better list of features in development, features that will be developed, changeLogs and other annotations the git hub project board at: [ML-CM Project Board ](https://github.com/Giacomo-Antonioli/Machine_Learning_Project/projects/1)
+
+To develop this project the following has been used:
+| Software kind     | Software version     |
+| ----------------  | -------------------  | 
+| Language          | Python 3.9.9         |
+| IDE               | Visual Studio Code   | 
+
+</p>
